Initialize `TradeableItem` with dataframe of core data
Pandas dataframes can be indexed on a specific list of data on
creation of the data frame by  using the `index` argument. Instead
 of initializing `tables` with an empty dataframe, as before, I
postpone initialization until after the core time series are
collected.

This implementation  means all the preprocessing and data verification
can be done before the table is populated. In return, the attribute
is robust to mistakes.

Secondly, this method will take advantage of indexing provided by
pandas.