Raise size mismatch error with `volume_series` (179 records)
PRICE data is requested from the OSRS GX API. VOLUME data is
webscraped directly from the listing page on runescape.com.

On the official page, both the PRICE and VOLUME graphs have
conccurrent timestamps. However, I observed that the requested
JSON includes 1 extra timestep; specifically, 1 day before the
earliest point on the offical listing graph. The resulting
behavior is the mismatch error.

Since the test in `tests.py` are explicitly checking that each
series is 180 records long and these tests passed on 13/6/20
I think there must be an official "update time" for one of
the series. So, I think it's best to leave the code unchanged
and check this hypothesis later.

If there is no obvious update time, then simply removing the
first record from the requested PRICE data will solve this
problem.