Format timestamps from epoch or ISO patterns
There is only a small time period where the timestamps for VOLUME
and PRICE series have concurrent timestamps. So, assuming that
timestamps from one can be used for another is only true for that
time period.

To make data collection robust to the the collection time (i.e. in
or out of the concurrent window) it's better to collect the *both*
sets of timestamps and merge wherever the sets are concurrent.

The tradeoff for robustness is that there will unavoidably be missing
data after merging. When the table is processed, the missing data
needs to be formatted.