Fix old school bond trade volume look up
Old school bonds are sold both on the Grand Exchange and Jagex's
online store. Since the Grand Exchange website only tracks trades
through the Grand Exchange, so there is no trade data available on
old school bonds.

When instantiating a `TradeableItem`, the trade volume is parsed
from the Grand Exchange webpage and added to a data frame of other
time series data. Since there is none to parse for old school bonds,
initializing `TradeableItem("old school bond")` raises an error-
specifically on formatting timestamps.

Old school bonds are, as of now, the only tradeable items without
a trade volume. So within `_initialize_table()`, there is a check
for which item is being initialized. If the item is not an old
school bond, the method works as before. if the item is an old
school bond, then only price data aggregate methods are called.