Run Fourier Forecast on updated data
About 20 days have passed since Fourier Forecast was run on bond
data. After updating the dataset, it is interesting to ask how
the new data would affect the forecast.

The most important influence new data can have is changing the
threshold value. So, the PSD was replotted to display the threshold.
From this plot, it is clear that 0.2e11 is still a valid value to
use, as the largest peaks are still greater than this value-albeit
only slightly.